Update to .net6, and DotVVM to 4.1.0-preview

+++ src/CheckBook.App/ViewModels/PaymentViewModel.cs
@@ -33,7 +33,7 @@ namespace CheckBook.App.ViewModels
 
 
         [Bind(Direction.ServerToClient)]
-        public decimal AmountDifference { get; set; }
+        public double AmountDifference { get; set; }
 
 
         [Bind(Direction.ServerToClient)]
