+++ src/CheckBook.DataAccess/Data/TransactionData.cs
@@ -8,7 +8,7 @@ namespace CheckBook.DataAccess.Data
 
         public int? UserId { get; set; }
 
-        public decimal? Amount { get; set; }
+        public double? Amount { get; set; }
 
         public string ImageUrl { get; set; }
