+++ src/CheckBook.DataAccess/Migrations/AppDbContextModelSnapshot.cs
@@ -61,9 +61,9 @@ namespace CheckBook.DataAccess.Migrations
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<decimal>("AmountNew");
+                    b.Property<double>("AmountNew");
 
-                    b.Property<decimal>("AmountOriginal");
+                    b.Property<double>("AmountOriginal");
 
                     b.Property<DateTime>("CreatedDate");
 
@@ -91,7 +91,7 @@ namespace CheckBook.DataAccess.Migrations
                     b.Property<int>("Id")
                         .ValueGeneratedOnAdd();
 
-                    b.Property<decimal>("Amount");
+                    b.Property<double>("Amount");
 
                     b.Property<int>("PaymentId");
