+++ src/CheckBook.DataAccess/Model/PaymentLog.cs
@@ -30,8 +30,8 @@ namespace CheckBook.DataAccess.Model
         [ForeignKey(nameof(UserId))]
         public virtual User User { get; set; }
 
-        public decimal AmountOriginal { get; set; }
+        public double AmountOriginal { get; set; }
 
-        public decimal AmountNew { get; set; }
+        public double AmountNew { get; set; }
     }
 }
